sunxi-gmac: Disable mac address based on md5(chip_id)
The chip_id is always 0 and thus the md5 sum for all Pine64 is the same
resulting in the same mac address. U-Boot has the chip_id so it is there
but the Kernel reads just zeros from the memory area used. We do not
care and simply disable this and just generate a random MAC address on
boot when none is set.